Revert "All movies page and graphs"
