--- .env.example
@@ -1,1 +1,0 @@
-VITE_MASTER_X_KEY=
