+++ src/App.tsx
@@ -5,11 +5,8 @@ import { PreLoader } from "./components/global/PreLoader";
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import { BrowseMovies } from "./pages/BrowseMovies";
 
-import { PublicRoute } from "./middleware/PublicRoute";
-
 const queryClient = new QueryClient();
 const Dashboard = lazy(() => import("@/pages/Dashboard"));
-const MovieDetail = lazy(() => import("@/pages/MovieDetail"));
 
 function App() {
   return (
@@ -17,31 +14,9 @@ function App() {
       <QueryClientProvider client={queryClient}>
         <BrowserRouter>
           <Routes>
-            <Route
-              path="/dashboard"
-              element={
-                <PublicRoute>
-                  <Dashboard />
-                </PublicRoute>
-              }
-            />
-            <Route
-              path="/"
-              element={
-                <PublicRoute>
-                  <BrowseMovies />
-                </PublicRoute>
-              }
-            />
-            <Route
-              path="/movies/:movieId"
-              element={
-                <PublicRoute>
-                  <MovieDetail />
-                </PublicRoute>
-              }
-            />
-            {/* <Route path="*" element={<Navigate to="/" replace />} /> */}
+            <Route path="/" element={<Navigate to="/dashboard" replace />} />
+            <Route path="/dashboard" element={<Dashboard />} />
+            <Route path="/browse" element={<BrowseMovies />} />
           </Routes>
         </BrowserRouter>
       </QueryClientProvider>
