+++ src/assets/data/SidebarData.ts
@@ -16,7 +16,7 @@ export const sidebarItems: SidebarItemTypes[] = [
     title: "Browse",
     icon: MdOutlineMovieFilter,
     comingSoon: true,
-    href: "/",
+    href: "/browse",
   },
 
   {
