+++ src/components/charts/BarGraph.tsx
@@ -9,8 +9,6 @@ import {
   LabelList,
   Legend,
   Brush,
-  CartesianGrid,
-  Rectangle,
 } from "recharts";
 
 interface BarGraphProps {
@@ -26,40 +24,46 @@ export const BarGraph = ({
   secondPlotDataKey,
 }: BarGraphProps) => {
   return (
-    <ResponsiveContainer width="100%" height="100%">
-      <BarChart
-        width={500}
-        height={300}
-        data={data}
-        margin={{
-          top: 5,
-          right: 30,
-          left: 20,
-          bottom: 5,
-        }}
-      >
-        <CartesianGrid strokeDasharray="3 3" />
+    <ResponsiveContainer
+      width="100%"
+      height="100%"
+      style={{ direction: "ltr" }}
+    >
+      <BarChart data={data} layout="vertical">
+        <XAxis type="number" />
+        <YAxis dataKey={dataKey} type="category" hide={true} />
+        <Tooltip />
+
         <Brush
           dataKey={dataKey}
           height={30}
           stroke="#8884d8"
           startIndex={0}
-          endIndex={6}
+          endIndex={3}
         />
-        <XAxis dataKey={dataKey} />
-        <YAxis />
-        <Tooltip />
-        <Legend />
+
         <Bar
           dataKey={firstPlotADataKey}
           fill="#8884d8"
-          activeBar={<Rectangle fill="pink" stroke="blue" />}
-        />
+          barSize={50}
+          stackId="a"
+          name={firstPlotADataKey}
+        >
+          <LabelList
+            dataKey={dataKey}
+            position="insideLeft"
+            fill="#ffffff"
+            style={{ fontWeight: "bold", fontSize: 12 }}
+          />
+        </Bar>
         <Bar
           dataKey={secondPlotDataKey}
           fill="#82ca9d"
-          activeBar={<Rectangle fill="gold" stroke="purple" />}
+          barSize={50}
+          stackId="a"
+          name={secondPlotDataKey}
         />
+        <Legend />
       </BarChart>
     </ResponsiveContainer>
   );
