+++ src/pages/Dashboard.tsx
@@ -1,26 +1,50 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
 import { Sidebar } from "../components/global/Sidebar";
-import { useSelector } from "react-redux";
+import { AppBar } from "../components/global/AppBar";
+import { useQuery } from "@tanstack/react-query";
+import { useDispatch } from "react-redux";
+import { setData } from "@/store/slices/MovieSlice";
 import { BarChartComponent } from "@/components/BarChartComponent";
-import { TopOscarWinsByGenre } from "@/components/TopOscarWinsByGenre";
-import { VerticalBarGraphComponent } from "@/components/VerticalBarGraphComponent";
-import { RootState } from "@/store";
-import { LoadingSpinner } from "@/components/global/LoadingSpinner";
 
 const Dashboard = () => {
-  const isLoading = useSelector(
-    (state: RootState) => state?.MoviesSlice?.isLoading
-  );
+  const [openSidebar, setOpenSidebar] = useState(false);
+  const dispatch = useDispatch();
+
+  const fetchMovies = async () => {
+    const response = await fetch(
+      "https://api.jsonbin.io/v3/b/67909b5bad19ca34f8f271d4/latest"
+    );
+    const data = await response.json();
+    return data;
+  };
+
+  const { isLoading, isError, data, error } = useQuery({
+    queryKey: ["movies"],
+    queryFn: fetchMovies,
+  });
+
+  useEffect(() => {
+    if (data) {
+      dispatch(setData(data.record));
+    }
+  }, [data]);
 
   return (
     <div className=" relative h-full">
-      <Sidebar />
-      <div className="xl:ml-[220px] grid grid-col-1 xl:grid-cols-2 gap-4">
-        {isLoading ? <LoadingSpinner /> : <VerticalBarGraphComponent />}
-        {isLoading ? <LoadingSpinner /> : <TopOscarWinsByGenre />}
-      </div>
-      <div className="xl:ml-[220px] ">
-        {isLoading ? <LoadingSpinner /> : <BarChartComponent />}
+      <section className="xl:absolute xl:flex hidden w-[200px] h-[100vh]  shadow-xl bg-white ">
+        <Sidebar />
+      </section>
+
+      <section className="flex xl:hidden">
+        <AppBar
+          openSidebar={openSidebar}
+          setOpenSidebar={setOpenSidebar}
+          pageName={"Home"}
+        />
+      </section>
+
+      <div className={` ml-[200px] bg-white text-black h-[100vh] px-6 pt-6`}>
+        {isLoading ? <p>Loading...</p> : <BarChartComponent />}
       </div>
     </div>
   );
