+++ src/store/slices/MovieSlice.ts
@@ -3,7 +3,6 @@ import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
   movies_data: moviesInitialData,
-  isLoading: true,
 };
 
 export const movieSlice = createSlice({
@@ -12,7 +11,6 @@ export const movieSlice = createSlice({
   reducers: {
     setData: (state, action) => {
       state.movies_data = action.payload;
-      state.isLoading = false;
     },
   },
 });
