+++ src/types/MovieDataTypes.ts
@@ -1,9 +1,4 @@
-export type GenreTypes = {
-  name: string;
-  value: number;
-};
 export type MovieDataTypes = {
-  id: string;
   title: string;
   year: string;
   genre: string[];
@@ -18,4 +13,17 @@ export type MovieDataTypes = {
   poster: string;
 };
 
-export const moviesInitialData: MovieDataTypes[] = [];
+export const moviesInitialData: MovieDataTypes = {
+  title: "",
+  year: "",
+  genre: [],
+  country: [],
+  imdb_rating: 0,
+  oscar_nominations: 0,
+  oscar_winning: 0,
+  cast: [],
+  language: [],
+  oscar_nominations_list: [],
+  oscar_winning_list: [],
+  poster: "",
+};
